+++ src/events/EventServiceProvider.js
@@ -22,10 +22,17 @@ import Dispatcher      from './services/Dispatcher';
 class EventServiceProvider extends ServiceProvider {
 
 	/**
+	 * @inheritdoc
+	 */
+	get name() {
+		return 'Node IoC - Events';
+	}
+
+	/**
 	 * Register the service provider.
 	 */
 	register() {
-		this.loadConfigFromFolder(__dirname, '..', 'config');
+		this.loadAndPublishConfig(this.app.formatPath(__dirname, 'config'));
 		this.bindEventDispatcher();
 	}