+++ src/support/services/Dumper.js
@@ -396,7 +396,7 @@ class Dumper extends checksTypes() {
 	 */
 	getLocation() {
 		return new Error().stack.split('\n').slice(1).filter((line) => {
-			return !line.includes(this.app.formatPath(__filename, '..', '..'));
+			return !line.includes(this.app.formatPath(__dirname, '..'));
 		}).shift().replace(/.*\((?<location>.*:\w+:\w+)\)/u, '$<location>');
 	}